requirements.txt
openai==1.12.0
anthropic==0.8.1
groq==0.4.1
google-cloud-aiplatform==1.38.1
cohere==4.39
python-dotenv==1.0.0
torch==2.1.2
transformers==4.37.2
numpy==1.24.3
pandas==2.1.4
streamlit==1.30.0
pytest==7.4.4
